requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
channels==4.2.0
channels_redis==4.2.1
Django==4.2.16
django-cors-headers==4.6.0
django-redis==5.4.0
djangorestframework==3.15.2
msgpack==1.1.0
pika==1.3.2
psutil==6.1.0
redis==5.2.0
sqlparse==0.5.2
